feat: enhance password reset functionality; refactor reset password form and improve error handling; remove unnecessary components
